Изменения в обработке данных о погоде и UI
Изменен тип свойства WindDirection с int на double в
WeatherJsonConverter.cs и Weather.cs для более точного
представления направления ветра. Добавлены опции
JsonSerializerOptions в WeatherService.cs для гибкой
десериализации JSON и добавления конвертеров
WeatherJsonConverter и CityJsonConverter. В
WeatherView.axaml добавлены элементы управления для
создания прокручиваемого списка, улучшая пользовательский
интерфейс.

+++ WeatherApp/Extensions/WeatherJsonConverter.cs
@@ -60,7 +60,7 @@ namespace WeatherApp.Extensions
                 if (wind.TryGetProperty("speed", out JsonElement speed))
                     weather.WindSpeed = speed.GetDouble();
                 if (wind.TryGetProperty("deg", out JsonElement deg))
-                    weather.WindDirection = deg.GetInt32();
+                    weather.WindDirection = deg.GetDouble();
             }
 
             if (root.TryGetProperty("clouds", out JsonElement clouds))
